refactor(wip): card routes
