--- src/routes/console/card/+layout.svelte
@@ -1,5 +1,0 @@
-<svelte:head>
-    <title>Card - Appwrite</title>
-</svelte:head>
-
-<slot />
