--- src/routes/console/card/+layout.ts
@@ -1,10 +1,0 @@
-import Header from './header.svelte';
-import Breadcrumbs from './breadcrumbs.svelte';
-import type { LayoutLoad } from './$types';
-
-export const load: LayoutLoad = async () => {
-    return {
-        header: Header,
-        breadcrumbs: Breadcrumbs
-    };
-};
