--- src/routes/console/card/breadcrumbs.svelte
@@ -1,12 +1,0 @@
-<script lang="ts">
-    import { Breadcrumbs } from '$lib/layout';
-
-    $: breadcrumbs = [
-        {
-            href: `/console`,
-            title: 'Console'
-        }
-    ];
-</script>
-
-<Breadcrumbs {breadcrumbs} />
